Improve layout  and hover highlighting
