+++ client/src/components/App/darkTheme.css
@@ -1,39 +1,2 @@
 @import url('https://fonts.googleapis.com/css2?family=Syne+Mono&display=swap');
 
-/* === BUTTONS === */
-
-.btn-nav.darkTheme {
-
-  color           : #eaeaea;
-  background-color: #1c1c1c;
-
-}
-
-.btn-nav.darkTheme.selected {
-  color           : black;
-  background-color: #eaeaea;
-  font-family     : inherit;
-  box-shadow      : inset 2px 2px #ffffff99, inset -2px -2px #00000099;
-}
-
-.btn-nav.darkTheme:hover {
-  background-color: rgb(97, 93, 93);
-}
-
-.btn-nav.darkTheme.selected:hover {
-  background-color: #eaeaea;
-}
-
-.btn-nav.darkTheme:focus {
-  /* remove annoying rounded blue outline when button is focused */
-  outline         : none;
-  /* change color when focused to resolve accessibility concerns */
-  background-color: #7b7b7bd3;
-}
-
-.btn-nav.darkTheme.selected:focus {
-  /* remove annoying rounded blue outline when button is focused */
-  outline         : none;
-  /* change color when focused to resolve accessibility concerns */
-  background-color: #eaeaea;
-}
