+++ client/src/components/Menu/Menu.css
@@ -1,25 +1,13 @@
 .Menu {
     grid-column      : 4/5;
     grid-row         : 2;
-    /* display       : block; */
-    /* display       : flex; */
-    /* flex-direction: row; */
+    background-color : var(--bg-dark);
     width            : max-content;
 }
 
-.greeting {
-    margin: auto 1rem;
-}
-
+.greeting,
 .menu-buttons {
     margin: auto 1rem;
-}
-
-.toggle {
-    background-color: #b34053;
-    width           : 10rem;
-    height          : 10rem;
-    font-size       : xx-large;
 }
 
 @media (max-width: 800px) {
