+++ client/src/components/Settings/Settings.tsx
@@ -3,22 +3,17 @@ import Admin from './Admin/Admin';
 import AutoSetup from './AutoSetup/AutoSetup';
 import General from './General/General';
 import Products from './Products/Products';
-import Investment from './Investment/Investment';
-import Reset from './Reset/Reset';
+import Portfolio from './Portfolio/Portfolio';
 import Feedback from './Feedback/Feedback';
-import History from './History/History';
-import SettingsNav from './SettingsNav/SettingsNav';
-import BulkDelete from './BulkDelete/BulkDelete';
-// import Draggable, { DraggerProps } from '../Draggable/Draggable';
 import Draggable from '../Draggable/Draggable';
 
 import './Settings.css'
 import { useUser } from '../../hooks/useUser';
 
 function Settings() {
-  const [settingsPage, setSettingsPage] = useState('general');
+  const [settingsPage, setSettingsPage] = useState('General');
   const [tips, setTips] = useState(false);
-  const { theme } = useUser();
+  const { user, theme } = useUser();
 
   const TipsCheck = useMemo(() => {
     return (
@@ -34,24 +29,35 @@ function Settings() {
     )
   }, [tips])
 
+  interface SettingsPages {
+    [key: string]: JSX.Element
+  }
+
+  const settingsPages: SettingsPages = useMemo(() => ({
+    'General': <General tips={tips} />,
+    'Portfolio': <Portfolio tips={tips} />,
+    'Auto Setup': <AutoSetup tips={tips} />,
+    'Products': <Products tips={tips} />,
+    'Feedback': <Feedback />,
+    'Admin': <Admin tips={tips} />
+  }), [tips])
+
   return (
     <Draggable className={`Settings ${theme}`} windowBarElements={[TipsCheck]}>
-      {/* <div> */}
-      <SettingsNav setSettingsPage={setSettingsPage} settingsPage={settingsPage} />
-      {
-        {
-          'general': <General tips={tips} />,
-          'products': <Products tips={tips} />,
-          'investment': <Investment tips={tips} />,
-          'autoSetup': <AutoSetup tips={tips} />,
-          'bulkDelete': <BulkDelete tips={tips} />,
-          'history': <History />,
-          'reset': <Reset tips={tips} />,
-          'feedback': <Feedback />,
-          'admin': <Admin tips={tips} />
-        }[settingsPage] // man this bit is clever. it's an object with a key that unlocks the desired component
-      }
-      {/* </div> */}
+      <div className="SettingsNav">
+        {Object.keys(settingsPages).map((page, i) => {
+          return (
+            page === 'Admin' && !user.admin
+              ? null
+              : <button
+                key={i}
+                className={`btn-nav ${theme} ${settingsPage === page && "selected"}`}
+                onClick={() => { setSettingsPage(page) }}
+              >{page}</button>
+          )
+        })}
+      </div>
+      {settingsPages[settingsPage]} {/* man this bit is clever. it's an object with a key that unlocks the desired component */}
     </Draggable>
   );
 }
