+++ client/src/components/SingleTrade/SingleTrade.css
@@ -1,128 +1,128 @@
 .Single-trade {
   display       : flex;
   flex-direction: row;
+
 }
 
 .Single-trade button {
   height: 1.7rem;
 }
 
-.overlay {
+.Single-trade div div {
+  margin: auto 0.5rem;
+  width : fit-content;
+}
+
+.Single-trade>div {
+  /* margin         : auto 0.5rem; */
+  width: fit-content;
+
   display        : flex;
   flex-direction : row;
   justify-content: space-around;
-  /* margin         : auto; */
-  /* flex-grow      : 1; */
   flex-wrap      : wrap;
   width          : 100%;
 }
 
-.overlay div {
-  margin: auto 0.5rem;
-  width        : fit-content;
-}
-
 @media screen and (max-width: 1230px) {
-  .overlay div {
-    flex-basis: 30%;
-    flex-grow : 1;
+  .Single-trade div div {
+    flex-basis     : 30%;
+    flex-grow      : 1;
     justify-content: space-evenly;
-    
+
     /* background-color: aquamarine; */
     /* flex-basis: 100%; */
   }
+
   /* any .overlay div on the right side should be right-justified */
-  .overlay div:nth-child(3n) {
+  .Single-trade div div:nth-child(3n) {
     text-align: right;
   }
-  .overlay div:nth-child(3n-1) {
+
+  .Single-trade div div:nth-child(3n-1) {
     text-align: center;
   }
 }
 
-
 @media screen and (max-width: 800px) {
-  .overlay div {
-    flex-basis: 33%;
-    /* flex-basis: 45%; */
-    flex-grow : 1;
-    justify-content: space-evenly;
-    min-width: fit-content;
+  .Single-trade div div {
+    flex-basis         : 33%;
+    /* flex-basis      : 45%; */
+    flex-grow          : 1;
+    justify-content    : space-evenly;
+    min-width          : fit-content;
     /* background-color: aquamarine; */
-    /* flex-basis: 100%; */
+    /* flex-basis      : 100%; */
   }
-  .overlay div:nth-child(3n) {
+
+  .Single-trade div div:nth-child(3n) {
     text-align: unset;
   }
-  .overlay div:nth-child(3n-1) {
+
+  .Single-trade div div:nth-child(3n-1) {
     text-align: unset;
   }
+
   /* any .overlay div on the right side should be right-justified */
-  .overlay div:nth-child(2n) {
+  .Single-trade div div:nth-child(2n) {
     text-align: right;
   }
 }
 
+:root {
+  --buy-color-light        : rgb(125, 209, 132);
+  --buy-color-light-shaded : rgb(96, 173, 103);
+  --sell-color-light       : rgb(199, 144, 112);
+  --sell-color-light-shaded: rgb(167, 119, 92);
+  --bg-color-dark          : rgb(65, 65, 65);
+  --bg-color-dark-shaded   : rgb(82, 82, 82);
+}
+
 .BUY {
-  border-top: solid 1px black;
-
-  height          : fit-content;
-  background-color: rgb(125, 209, 132);
-}
-
-.BUY.darkTheme {
-  border-top         : solid 1px black;
-  height             : fit-content;
-  /* background-color: rgb(125, 209, 132); */
-  background-color   : rgb(65, 65, 65);
-  color              : rgb(125, 209, 132);
-}
-
-.BUY:nth-child(2n) {
-  background-color: rgb(96, 173, 103);
-}
-
-.BUY.darkTheme:nth-child(2n) {
-  /* background-color: rgb(96, 173, 103); */
-  background-color: rgb(82, 82, 82);
+  border-top      : solid 1px black;
+  background-color: var(--buy-color-light);
 }
 
 .SELL {
   border-bottom   : black 1px solid;
-  background-color: rgb(167, 119, 92);
+  background-color: var(--sell-color-light);
+}
+
+.BUY.darkTheme {
+  border-top: solid 1px black;
+  color     : var(--buy-color-light);
 }
 
 .SELL.darkTheme {
-  border-bottom      : black 1px solid;
-  /* background-color: rgb(167, 119, 92); */
-  background-color   : rgb(65, 65, 65);
-  /* color           : rgb(167, 119, 92); */
-  color              : rgb(199, 144, 112);
-  /* color           : rgb(255, 255, 255); */
+  border-bottom: black 1px solid;
+  color        : var(--sell-color-light);
 }
 
-.SELL:nth-child(2n) {
-  background-color: rgb(199, 144, 112);
+/* set background when not hover */
+.BUY.darkTheme:not(:hover) {
+  background-color: var(--bg-color-dark);
 }
 
-.SELL.darkTheme:nth-child(2n) {
-  background-color   : rgb(82, 82, 82);
-  /* background-color: rgb(199, 144, 112); */
-  /* color           : rgb(199, 144, 112); */
+.SELL.darkTheme:not(:hover) {
+  background-color: var(--bg-color-dark);
 }
 
-
-.overlay:hover {
-  /* z-index            : 9; */
-  /* width              : 100%; */
-  /* height             : 100%; */
-  /* position        : absolute; */
-  background-color      : #2cb7da46;
-  /* opacity            : 0.5; */
-  /* filter             : brightness(1.1); */
-  /* border-radius      : 30px; */
+.BUY:nth-child(2n):not(:hover) {
+  background-color: var(--buy-color-light-shaded);
 }
 
+.SELL:nth-child(2n):not(:hover) {
+  background-color: var(--sell-color-light-shaded);
+}
+
+.BUY.darkTheme:nth-child(2n):not(:hover) {
+  /* background-color: rgb(96, 173, 103); */
+  background-color: rgb(82, 82, 82);
+}
+
+.SELL.darkTheme:nth-child(2n):not(:hover) {
+  background-color: rgb(82, 82, 82);
+}
 
 .deleting,
 .BUY .kill-button,
